+++ src/main/java/in/koreatech/koin/domain/community/articles/controller/CommunityController.java
@@ -1,30 +1,21 @@
 package in.koreatech.koin.domain.community.articles.controller;
 
-import static in.koreatech.koin.domain.user.model.UserType.STUDENT;
-
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import in.koreatech.koin.domain.community.articles.dto.ArticleHotKeywordResponse;
-import in.koreatech.koin.domain.community.keywords.dto.ArticleKeywordCreateRequest;
-import in.koreatech.koin.domain.community.keywords.dto.ArticleKeywordResponse;
 import in.koreatech.koin.domain.community.articles.dto.ArticleResponse;
 import in.koreatech.koin.domain.community.articles.dto.ArticlesResponse;
 import in.koreatech.koin.domain.community.articles.dto.HotArticleItemResponse;
 import in.koreatech.koin.domain.community.articles.service.CommunityService;
-import in.koreatech.koin.global.auth.Auth;
 import in.koreatech.koin.global.auth.UserId;
 import in.koreatech.koin.global.ipaddress.IpAddress;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
 @RestController
