+++ src/main/java/in/koreatech/koin/domain/community/keyword/controller/KeywordApi.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.community.keywords.controller;
+package in.koreatech.koin.domain.community.keyword.controller;
 
 import static in.koreatech.koin.domain.user.model.UserType.STUDENT;
 
@@ -9,11 +9,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
-import in.koreatech.koin.domain.community.keywords.dto.ArticleKeywordCreateRequest;
-import in.koreatech.koin.domain.community.keywords.dto.ArticleKeywordResponse;
-import in.koreatech.koin.domain.community.keywords.dto.ArticleKeywordsResponse;
-import in.koreatech.koin.domain.community.keywords.dto.ArticleKeywordsSuggestionResponse;
-import in.koreatech.koin.domain.community.keywords.dto.KeywordNotificationRequest;
+import in.koreatech.koin.domain.community.keyword.dto.ArticleKeywordCreateRequest;
+import in.koreatech.koin.domain.community.keyword.dto.ArticleKeywordResponse;
+import in.koreatech.koin.domain.community.keyword.dto.ArticleKeywordsResponse;
+import in.koreatech.koin.domain.community.keyword.dto.ArticleKeywordsSuggestionResponse;
+import in.koreatech.koin.domain.community.keyword.dto.KeywordNotificationRequest;
 import in.koreatech.koin.global.auth.Auth;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
