+++ src/test/java/in/koreatech/koin/AcceptanceTest.java
@@ -22,11 +22,7 @@ import org.testcontainers.utility.DockerImageName;
 import in.koreatech.koin.config.TestJpaConfiguration;
 import in.koreatech.koin.config.TestRedisConfiguration;
 import in.koreatech.koin.config.TestTimeConfig;
-import in.koreatech.koin.domain.bus.util.CityBusClient;
-import in.koreatech.koin.domain.bus.util.CityBusRouteClient;
-import in.koreatech.koin.domain.bus.util.PublicExpressBusClient;
-import in.koreatech.koin.domain.bus.util.TmoneyExpressBusClient;
-import in.koreatech.koin.domain.community.keywords.model.ArticleKeywordEventListener;
+import in.koreatech.koin.domain.community.keyword.model.ArticleKeywordEventListener;
 import in.koreatech.koin.domain.coop.model.CoopEventListener;
 import in.koreatech.koin.domain.owner.model.OwnerEventListener;
 import in.koreatech.koin.domain.shop.model.ShopEventListener;
