+++ src/test/java/in/koreatech/koin/acceptance/KeywordApiTest.java
@@ -15,11 +15,11 @@ import org.springframework.http.HttpStatus;
 import in.koreatech.koin.AcceptanceTest;
 import in.koreatech.koin.domain.community.articles.model.Article;
 import in.koreatech.koin.domain.community.articles.model.Board;
-import in.koreatech.koin.domain.community.keywords.model.ArticleKeywordSuggestCache;
-import in.koreatech.koin.domain.community.keywords.model.ArticleKeywordUserMap;
-import in.koreatech.koin.domain.community.keywords.repository.ArticleKeywordRepository;
-import in.koreatech.koin.domain.community.keywords.repository.ArticleKeywordSuggestRepository;
-import in.koreatech.koin.domain.community.keywords.repository.ArticleKeywordUserMapRepository;
+import in.koreatech.koin.domain.community.keyword.model.ArticleKeywordSuggestCache;
+import in.koreatech.koin.domain.community.keyword.model.ArticleKeywordUserMap;
+import in.koreatech.koin.domain.community.keyword.repository.ArticleKeywordRepository;
+import in.koreatech.koin.domain.community.keyword.repository.ArticleKeywordSuggestRepository;
+import in.koreatech.koin.domain.community.keyword.repository.ArticleKeywordUserMapRepository;
 import in.koreatech.koin.domain.user.model.Student;
 import in.koreatech.koin.fixture.ArticleFixture;
 import in.koreatech.koin.fixture.BoardFixture;
